Restructure commands (#96)
This implements the design we decided on earlier here (https://github.com/qlik-oss/corectl/blob/command-restructure-design/command-design.md) with some small diffs.

The `get` commands are pure and will only get what is in the app, and not set anything.

Added remove commands so you can not only set objects/measures/dimensions but remove them too.



When set:ing something, measures, for example, the measures are picked up from BOTH the command line and the config file. This is the current behavior and has not been changed but perhaps it should? (But in a different PR then)


Structurally I have made some clean-ups in the code broken out a few methods but the focus has been on implementing the new and cleaner API. And more refactoring of the code could be done in the future.

Bidning flags to viper is now done in the `PersistentPreRun` functions since doing it in the `init` method meant that one flag could only be bound to one command.

I think the easiest way to test this is to clone the branch and build the tool and use the new API and check if it feels nice.

This Closes #95, closes #66, closes #86, closes #72, closes #75, closes #73